// src/app/modules/auth/models/permissions.model.ts
export class PermissionsModel {
  Dashboard: boolean = false;
  EmployeeMaster: boolean = false;
  StockistMaster: boolean = false;
  CartingMaster: boolean = false;
  CourierMaster: boolean = false;
  BranchMaster: boolean = false;
  CompanyMaster: boolean = false;
  StockistTransporterMapping: boolean = false;
  TransporterMaster: boolean = false;
  StockistBranchRelation: boolean = false;
  StockistCompanyRelation: boolean = false;
  GeneralMaster: boolean = false;
  EmailConfig: boolean = false;
  PicklistOperation: boolean = false;
  PicklistAdd: boolean = false;
  PicklistVerify: boolean = false;
  PicklistAllot: boolean = false;
  ReAllotPicklist: boolean = false;
  ImportInvData: boolean = false;
  InvCancelList: boolean = false;
  ReadyToDispatch: boolean = false;
  AssignTransportMode: boolean = false;
  ImportLRData: boolean = false;
  PrintSticker: boolean = false;
  ChqRegister: boolean = false;
  ImportOS: boolean = false;
  ImportChqDeposit: boolean = false;
  ChqSummRpt: boolean = false;
  ChqSummMonthlyRpt: boolean = false;
  ResolveConcernPL: boolean = false;
  ResolveConcernINV: boolean = false;
  PriotiseINV: boolean = false;
  AssignTransportModeEdit: boolean = false;
  AppConfig: boolean = false;
  ImportTransitReport: boolean = false;
  ApproveVehicleIssue: boolean = false;
  InsuranceClaim: boolean = false;
  ApprovalClaim: boolean = false;
  LRReceivedList: boolean = false;
  ResolveClaimConcern: boolean = false;
  ImportSRS: boolean = false;
  CorrectionRequiredList: boolean = false;
  SRSPendingCNList: boolean = false;
  ImportCreditNote: boolean = false;
  UploadDestructionCertificate: boolean = false;
  DestructionCertificateList: boolean = false;
  AppConfiguration: boolean = false;
  StockTransferAdd: boolean = false;
  CityMaster: boolean = false;
  ThresholdValueMaster: boolean = false;
  ChecklistMaster: boolean = false;
  OtherCNFMaster: boolean = false;
  ImportDepositedCheque: boolean = false;
  ChequeRegisterSummaryReport: boolean = false;
  RaisedConcernList: boolean = false;
  VersionDetails: boolean = false;
  BranchCompanyRelationMaster: boolean = false;
  ExpenseRegister: boolean = false;
  ReimbursmentInvoice: boolean = false;
  ComissionInvoice: boolean = false;
  GatepassbillSummary: boolean = false;
  CheckInvoice: boolean = false;
  VendorMaster: boolean = false;
  TaxMaster: boolean = false;
  HeadMaster: boolean = false;
  TransportParentMaster: boolean = false;
  TransportParentMapping: boolean = false;
  CourierParentMaster: boolean = false;
  CourierParentMapping: boolean = false;
  CompanyVendorMapping: boolean = false;
  ChequeStatusReport: boolean = false;
  VendorBranchMapping: boolean = false;
  VehicleChecklistForImg: boolean = false;
  VerifyStockistData:boolean = false;

  // FOR ALL DASHBORD PERMISSIONS
  DashBoardOrderReturn: boolean = false;
  DashBoardOrderDispatch: boolean = false;
  DashBoradChequeAcc: boolean = false;
  DashBoardInventoyInward: boolean = false;
  DashBoradStockTrans: boolean = false;
  DashMiniORPendingSRS: boolean = false;
  DashORFOROperator: boolean = false;
  DashCheackAccForOprator: boolean = false;
  DashOrdDisForOperator: boolean = false;
  DashOrdDisForSupervisor: boolean = false;
  DashBoardORForSupervisor: boolean = false;
  DshBranchDRP: boolean = false;
  DshCompanyDRP: boolean = false;
}